requirements.txt
Pillow~=8.3.1
pytz~=2021.1
numpy~=1.21.0
setuptools~=57.1.0
scipy~=1.7.0
matplotlib~=3.4.2
joblib~=1.0.1
pandas~=1.3.0
openpyxl~=3.0.7
python-dateutil~=2.8.1
SQLAlchemy~=1.4.20
six~=1.16.0
plotly~=5.1.0
tenacity~=8.0.1
sklearn~=0.0
scikit-learn~=0.24.2
threadpoolctl~=2.2.0
catboost~=0.26
greenlet~=1.1.0
cycler~=0.10.0
pyparsing~=2.4.7
kiwisolver~=1.3.1
zipp~=3.5.0
cx-Oracle==8.2.1
pyyaml == 5.4.1
lasio == 0.29
transliterate == 1.10.2
